ml-services/crowd_counting/app.py
from flask import Flask, request, jsonify
from flask_cors import CORS
import base64
import numpy as np
from io import BytesIO
from PIL import Image
import cv2
import time
from ultralytics import YOLO

app = Flask(__name__)
CORS(app)

# Models
yolo_model = None
# MCNN would be implemented separately for high-density crowds

def load_models():
    """Load YOLO model for face detection"""
    global yolo_model
    
    try:
        # Load YOLOv8 model (can be trained specifically for faces)
        # For now, using general object detection
        yolo_model = YOLO('yolov8n.pt')  # Nano model for speed
        print("YOLO model loaded successfully")
    except Exception as e:
        print(f"Error loading YOLO model: {e}")

def generate_density_map(image, detections):
    """Generate density heatmap from detections"""
    height, width = image.shape[:2]
    density_map = np.zeros((height, width), dtype=np.float32)
    
    # Create Gaussian kernel
    sigma = 15
    kernel_size = sigma * 3
    
    for detection in detections:
        x, y, w, h = detection
        cx, cy = int(x + w/2), int(y + h/2)
        
        # Add Gaussian blob at face center
        y1 = max(0, cy - kernel_size)
        y2 = min(height, cy + kernel_size)
        x1 = max(0, cx - kernel_size)
        x2 = min(width, cx + kernel_size)
        
        for i in range(y1, y2):
            for j in range(x1, x2):
                dist = np.sqrt((i - cy)**2 + (j - cx)**2)
                density_map[i, j] += np.exp(-(dist**2) / (2 * sigma**2))
    
    # Normalize and convert to heatmap
    if density_map.max() > 0:
        density_map = (density_map / density_map.max() * 255).astype(np.uint8)
    
    # Apply colormap
    heatmap = cv2.applyColorMap(density_map, cv2.COLORMAP_JET)
    
    # Blend with original image
    blended = cv2.addWeighted(image, 0.6, heatmap, 0.4, 0)
    
    return blended

@app.route('/health', methods=['GET'])
def health():
    return jsonify({
        'status': 'ok',
        'service': 'crowd_counting',
        'model_loaded': yolo_model is not None
    })

@app.route('/count', methods=['POST'])
def count():
    try:
        start_time = time.time()
        
        data = request.json
        image_data = data.get('image')
        
        if not image_data:
            return jsonify({'error': 'No image provided'}), 400
        
        # Decode base64 image
        image_bytes = base64.b64decode(image_data.split(',')[1] if ',' in image_data else image_data)
        image = Image.open(BytesIO(image_bytes))
        image_np = np.array(image)
        
        face_count = 0
        detections = []
        
        if yolo_model is not None:
            # Run YOLO detection
            results = yolo_model(image_np, conf=0.25)
            
            # Count persons (class 0 in COCO dataset)
            # In production, use a face-specific YOLO model
            for result in results:
                boxes = result.boxes
                for box in boxes:
                    cls = int(box.cls[0])
                    if cls == 0:  # Person class
                        face_count += 1
                        x1, y1, x2, y2 = box.xyxy[0].cpu().numpy()
                        detections.append([int(x1), int(y1), int(x2-x1), int(y2-y1)])
        else:
            # Placeholder
            face_count = 150
            detections = [[100, 100, 50, 50]]
        
        # Generate density map
        density_map_img = generate_density_map(image_np, detections)
        
        # Convert density map to base64
        _, buffer = cv2.imencode('.jpg', density_map_img)
        density_map_base64 = base64.b64encode(buffer).decode('utf-8')
        
        processing_time = (time.time() - start_time) * 1000
        
        # Estimate accuracy based on crowd density
        density = 'low' if face_count < 50 else 'medium' if face_count < 200 else 'high'
        accuracy = 0.95 if density == 'low' else 0.90 if density == 'medium' else 0.85
        
        return jsonify({
            'count': face_count,
            'density_map': density_map_base64,
            'confidence': accuracy,
            'processing_time': processing_time,
            'crowd_density': density
        })
    except Exception as e:
        return jsonify({'error': str(e)}), 500

if __name__ == '__main__':
    load_models()
    app.run(host='0.0.0.0', port=5003, debug=True)
