ml-services/crowd_counting/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.0.232
opencv-python==4.9.0.80
numpy==1.24.3
pillow==10.1.0
torch==2.1.2
torchvision==0.16.2
